did modifications after changing to package structure
