+++ run.py
@@ -1,0 +1,5 @@
+from flaskblog import app
+
+# Makes sure app runs in debug
+if __name__ == '__main__':
+    app.run(debug = True)
